packages/linux-lava: hack further to only overlay for linux-lava specifically
wowie wowzers, this is a terrible hack. All this is needed only because
I couldn't for the life of me find a way to override rust-bindgen from
buildLinux.

Simply including the overlay would require rebuilding a crapton of
things that depend on rust-bindgen, so well, this'll have to do for now.